requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
asgiref==3.7.2
asttokens==2.4.1
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
colorama==0.4.6
cryptography==41.0.4
decorator==5.1.1
Django==3.2.22
django-auditlog==2.3.0
django-cors-headers==4.3.1
django-filter==23.3
django-js-asset==2.1.0
django-mptt==0.14.0
djangorestframework==3.14.0
docopt==0.6.2
drf-excel==2.4.0
drf-flex-fields==1.0.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
executing==2.0.0
filelock==3.12.4
fsspec==2023.10.0
google-api-core==2.12.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
huggingface-hub==0.20.2
idna==3.4
inflect==7.0.0
inflection==0.5.1
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
kaldi-io==0.9.8
lazy_loader==0.3
ldap3==2.9.1
librosa==0.10.1
llvmlite==0.41.1
Markdown==3.5
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.7
NEMO==5.0.0
nemo_toolkit @ git+https://github.com/NVIDIA/NeMo.git@177a67fa4a43ed6f7c4fb39a66c0a2998a27fef1
networkx==3.2
num2words==0.5.13
numba==0.58.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.15.0
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
pooch==1.8.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pymodbus==3.3.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.18.2
ruamel.yaml.clib==0.2.8
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
soundfile==0.12.1
soupsieve==2.5
sox==1.4.1
soxr==0.3.7
SpeechRecognition==3.10.0
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.0
torch-stft==0.1.4
torchvision==0.16.0
tqdm==4.66.1
traitlets==5.12.0
transformers==4.36.2
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.0.7
user-agent==0.1.10
wcwidth==0.2.8
Werkzeug==3.0.1
wget==3.2
wrapt==1.15.0
